refactor: extract layoutKey helper to reuse in hooks
