+++ src/components/plugin-wrapper/hooks/use-line-list-analytics-data.ts
@@ -27,8 +27,8 @@ import type {
     MetadataInputItem,
     UserOrgUnitMetadataItem,
 } from '@types'
-import { useCallback, useState } from 'react'
-import { getAnalyticsEndpoint } from './query-tools-common'
+import { useCallback, useMemo, useState } from 'react'
+import { computeLayoutKey, getAnalyticsEndpoint } from './query-tools-common'
 import { getAdaptedVisualization } from './query-tools-line-list'
 
 const lookupOptionSetOptionMetadata = (optionSetId, code, metaDataItems) => {
@@ -317,7 +317,7 @@ type AnalyticsDataState = {
 type UseAnalyticsDataResult = [FetchAnalyticsDataFn, AnalyticsDataState]
 
 const useLineListAnalyticsData = (
-    layoutKey: string
+    visualization: CurrentVisualization
 ): UseAnalyticsDataResult => {
     const dataEngine = useDataEngine()
     const metadataStore = useMetadataStore()
@@ -328,6 +328,11 @@ const useLineListAnalyticsData = (
         error: undefined,
         data: null,
     })
+
+    const layoutKey = useMemo(
+        () => computeLayoutKey(visualization),
+        [visualization]
+    )
 
     const [prevLayoutKey, setPrevLayoutKey] = useState(layoutKey)
     if (prevLayoutKey !== layoutKey) {
