+++ src/components/plugin-wrapper/line-list-plugin.tsx
@@ -31,20 +31,8 @@ export const LineListPlugin: FC<LineListPluginProps> = ({
     onDataSorted,
     onResponseReceived,
 }) => {
-    const layoutKey = useMemo(
-        () =>
-            `${visualization.outputType}|${[
-                ...visualization.columns,
-                ...visualization.rows,
-                ...visualization.filters,
-            ]
-                .map((d) => d.dimension)
-                .join('|')}`,
-        [visualization]
-    )
-
     const [fetchAnalyticsData, { data, isFetching }] =
-        useLineListAnalyticsData(layoutKey)
+        useLineListAnalyticsData(visualization)
 
     // null indicates no custom sorting has been applied
     // undefined cannot be used because that is a valid value to indicate "remove sorting"
