+++ src/components/plugin-wrapper/pivot-table-plugin.tsx
@@ -61,19 +61,8 @@ export const PivotTablePlugin: FC<PivotTablePluginProps> = ({
         [visualization]
     )
 
-    const layoutKey = useMemo(
-        () =>
-            `${visualization.outputType}|${[
-                ...visualization.columns,
-                ...visualization.rows,
-                ...visualization.filters,
-            ]
-                .map((d) => d.dimension)
-                .join('|')}`,
-        [visualization]
-    )
-
-    const [fetchAnalyticsData, { data }] = usePivotTableAnalyticsData(layoutKey)
+    const [fetchAnalyticsData, { data }] =
+        usePivotTableAnalyticsData(visualization)
 
     // TODO: implement onDataSorted and any other function/callback that cannot rely on the Redux store
